pyproject.toml
[tool.poetry]
name = "eventcatalog-api"
version = "0.1.0"
description = "Event Catalog GraphQL API"
authors = ["Kendy C. <ken.cj4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
strawberry-graphql = {extras = ["asgi"], version = "^0.214.0"}
fastapi = "^0.104.1"
jinja2 = "^3.1.2"
faker = "^22.2.0"
neo4j = "^5.15.0"
neomodel = "^5.2.1"
pydantic-settings = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^23.11"
mypy = "^1.7.0"
ruff = "^0.0.47"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = [
    "F401",
    "F403",
    "E501",
]